requirements.txt
fastapi
uvicorn
openai
gspread
google-auth
apscheduler
razorpay
python-dotenv
httpx
pytz
